patch: move listFiles
